+++ src/LionWeb.Protocol.Delta/Globals.cs
@@ -34,3 +34,6 @@ global using EventSequenceNumber = long;
 global using SemanticPropertyValue = object;
 global using ErrorCode = string;
 global using SharedKeyedMap = System.Collections.Generic.Dictionary<LionWeb.Core.Serialization.MetaPointer, LionWeb.Core.M3.IKeyed>;
+global using IndexOffset = int;
+global using ContinuedChunkCompleted = bool;
+global using ContinuedChunkSequenceNumber = long;