+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres _loosely_ to [Semantic Versioning](https://semver.org/s
 * Adjusted to latest spec changes:
   * `AdditionalInfo`: `Distribute` flag, Dictionary instead of array of `AdditionalInfoData`
   * `ReconnectRequest`: added `DeltaProtocolVersion`, `ClientId`, `RepositoryId`
+  * Introduced `IDeltaSplittable` and `IDeltaContinued` types (no functionality yet)
 ### Removed
 ### Deprecated
 ### Security